% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{kf_decomp_filter}
\alias{kf_decomp_filter}
\title{Kalman filter an estimated model from kf_decomp_estim}
\usage{
kf_decomp_filter(y, model, plot = F, select = NULL)
}
\arguments{
\item{y}{Univariate time series of data values. May also be a 2 column data frame containing a date column.}

\item{model}{Structural time series model estimated using kf_decomp_estim.}

\item{plot}{Logial, whether to plot the output or not.}

\item{select}{If NULL, then use the model selected by kf_dceomp_estim; otherwise, use the one specified with select}
}
\value{
List of data tables containing the filtered and smoothed series.
}
\description{
Kalman filter an estimated model from kf_decomp_estim
}
\examples{
kf_decomp_filter(y = DT[, c("date", "y")], model = model)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
