% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{tcs_boxcox}
\alias{tcs_boxcox}
\title{Box-Cox Transformation}
\usage{
tcs_boxcox(y, test = T, reverse = F, lambda = NULL, plot = F)
}
\arguments{
\item{y}{A univariate time series}

\item{test}{Whether to perform the test to select the best power transformation parameter}

\item{reverse}{Wheter to perform the reverse transformation}

\item{lambda}{The Box-Cox power parameter}
}
\description{
Fits a Box-Cox transformation and transformations the data
}
