% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{tcs_build_dates}
\alias{tcs_build_dates}
\title{Build the date sequence as a Date type}
\usage{
tcs_build_dates(y)
}
\arguments{
\item{y}{an object created from tcs_detect_freq}
}
\description{
Build the date sequence as a Date type
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
