% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{tcs_detect_decomp}
\alias{tcs_detect_decomp}
\title{Detect cycle and seasonality from the data}
\usage{
tcs_detect_decomp(y, freq, level = 0.01, method = "ARIMA", n.sim = 100)
}
\arguments{
\item{y}{Univariate time series of data values. May also be a 2 column data frame containing a date column.}

\item{freq}{Seasonality of the data (1, 4, 12, 52, 365)}

\item{method}{Method for wavelet analysis comparison ("white.nose", "shuffle", "Fourier.rand", "AR", "ARIMA"). Default is "ARIMA".}

\item{n.sim}{Number of simulations for wavelet analysis. Default is 100}
}
\value{
List giving the decomposition and wavelet
}
\description{
Detect cycle and seasonality from the data
}
