% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{tcs_detect_freq}
\alias{tcs_detect_freq}
\title{Detect frequency and dates from the data}
\usage{
tcs_detect_freq(y, freq = NULL)
}
\arguments{
\item{y}{Univariate time series of data values. May also be a 2 column data frame containing a date column.}

\item{freq}{Initial setting for the frequency detection}
}
\value{
List giving the dates and frequency of the data
}
\description{
Detect frequency and dates from the data
}
