% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{tcs_ssm}
\alias{tcs_ssm}
\title{State space model
Creates a state space model in list form
yt = Ht * Bt + e_t
Bt = Ft * B_{t-1} + u_t}
\usage{
tcs_ssm(par = NULL, yt = NULL, freq = NULL, decomp = NULL,
  trend_spec = NULL, init = NULL, model = NULL, full_seas_freq = F)
}
\arguments{
\item{par}{Vector of named parameter values}

\item{yt}{Univariate time series of data values}

\item{freq}{Seasonality of the data (1, 4, 12, 52, 365)}

\item{decomp}{Decomposition model ("tend-cycle-seasonal", "trend-seasonal", "trend-cycle", "trend-noise")}

\item{trend_spec}{Trend specification (NULL, "rw", "rwd", "2rw"). The default is NULL which will choose the best of all specifications based on the maximum likielhood. 
"rw" is the random walk trend. "rwd" is the random walk with random walk drift trend. "2rw" is a 2nd order random walk trend.}

\item{init}{Initial state values for the Kalman filter}
}
\value{
List of space space matrices
}
\description{
State space model
Creates a state space model in list form
yt = Ht * Bt + e_t
Bt = Ft * B_{t-1} + u_t
}
\examples{
tcs_ssm(par, y, freq, decomp, trend_spec, init)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
