% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{SSmodel}
\alias{SSmodel}
\title{State space model
Creates a state space model in list form
yt = Ht %*% Bt + e_t
Bt = Ft %*% B_{t=1} + u_t}
\usage{
SSmodel(par, yt, freq, decomp, int_order, trend, init = NULL)
}
\arguments{
\item{par}{Vector of named parameter values}

\item{yt}{Univariate time series of data values}

\item{freq}{Seasonality of the data (1, 4, 12, 53, 365)}

\item{decomp}{Decomposition model ("tend-cycle-seasonal", "trend-seasonal", "trend-cycle", "trend-noise")}

\item{int_order}{Order of integration (0, 1, 2)}

\item{init}{Initial state values for the Kalman filter}

\item{tend}{Trend specification ("", "rw", "hp")}
}
\value{
List of space space matrices
}
\description{
State space model
Creates a state space model in list form
yt = Ht %*% Bt + e_t
Bt = Ft %*% B_{t=1} + u_t
}
\examples{
SSmodel(par, y, freq, decomp, int_order, trend, init)
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
