% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_functions.R
\name{tcs_decomp_estim}
\alias{tcs_decomp_estim}
\title{Trend cycle seasonal decomposition using the Kalman filter
Estimates a structural time series model using the Kalman filter and maximum likelihood
The seasonal and cycle components are assumed to be of a trigonometric form.
The function checks 4 trend specifications to decompose a univariate time series
into trend, cycle, and/or seasonal components plus noise. The function automatically
detects the frequency and checks for a seasonal and cycle component if the user does not specify
the frequency or decomposition model. This can be turned off by setting freq or specifying decomp.}
\usage{
tcs_decomp_estim(y, freq = NULL, decomp = NULL, int_order = NULL,
  blend = F, level = 0.01, ur_test = "adf", ur_type = "level",
  optim_methods = c("BFGS", "CG", "NM"), maxit = 1000,
  maxtrials = 10)
}
\arguments{
\item{y}{Univariate time series of data values. May also be a 2 column data frame containing a date column.}

\item{freq}{Seasonality of the data (1, 4, 12, 53, 365)}

\item{decomp}{Decomposition model ("tend-cycle-seasonal", "trend-seasonal", "trend-cycle", "trend-noise")}

\item{int_order}{Order of integration (0, 1, 2)}

\item{blend}{Blend the model specifications using a minimum mean squared error weighted sum}

\item{level}{Significance level of statistical tests (0.01, 0.05, 0.1)}

\item{ur_test}{Unit root test name ("kpss", "adf", "pp")}

\item{ur_type}{Unit root test deterministi type ("level", "trend")}

\item{optim_methods}{Vector of 1 to 3 optimization methods in order of preference ("NR", "BFGS", "CG", "BHHH", or "SANN")}

\item{maxit}{Maximum number of iterations for the optimization}

\item{maxtrials}{Maximum number of optimization trials to get convergence}
}
\value{
List of estimation values including coefficients, convergence code, datea frequency, decomposition used, and trend specification selected.
}
\description{
State space model for decomposition
Yt = T_t + C_t + S_t + e_t, e_t ~ N(0, sig_e^2)
Y is the data
T is the trend component
C is the cycle component
S is the seasonal component
e is the observation error
}
\examples{
tcs_decomp_estim(y = DT[, c("date", "y")])
}
\author{
Alex Hubbard (hubbard.alex@gmail.com)
}
